Update functions for Admin, Manager, fixed interface of both , fixed users, changed class of admin manager and user, fiexd para of vehicles ( int to double )

+++ IVehicle.java
@@ -3,9 +3,9 @@ import java.util.Map;
 public interface IVehicle {
     String getVehicleID(); // Get vehicle ID
     String getName(); // Get vehicle name
-    int getCurrentFuel(); // Get current fuel
-    int getFuelCapacity(); // Get vehicle fuel capacity
-    int getCarryCapacity(); // Get vehicle carry capacity
+    double getCurrentFuel(); // Get current fuel
+    double getFuelCapacity(); // Get vehicle fuel capacity
+    double getCarryCapacity(); // Get vehicle carry capacity
     int getTotalContainers(); // Get total containers
     Port getCurretPort(); // Get the current port
     Map<Container, Integer> getContainerCounts();
